+++ packages/lib/test/action/applySet.test.ts
@@ -59,6 +59,7 @@ test("applySet", () => {
           "rootContext": [Circular],
           "target": P {
             "$": Object {
+              "$modelId": "id-1",
               "ch": undefined,
               "obj": undefined,
               "y": 5,
@@ -81,6 +82,7 @@ test("applySet", () => {
           "rootContext": [Circular],
           "target": P {
             "$": Object {
+              "$modelId": "id-1",
               "ch": undefined,
               "obj": undefined,
               "y": 5,
@@ -119,6 +121,7 @@ test("applySet", () => {
             "ch",
             P {
               "$": Object {
+                "$modelId": "id-2",
                 "ch": undefined,
                 "obj": undefined,
                 "y": 1,
@@ -131,8 +134,10 @@ test("applySet", () => {
           "rootContext": [Circular],
           "target": P {
             "$": Object {
+              "$modelId": "id-1",
               "ch": P {
                 "$": Object {
+                  "$modelId": "id-2",
                   "ch": undefined,
                   "obj": undefined,
                   "y": 1,
@@ -155,6 +160,7 @@ test("applySet", () => {
             "ch",
             P {
               "$": Object {
+                "$modelId": "id-2",
                 "ch": undefined,
                 "obj": undefined,
                 "y": 1,
@@ -167,8 +173,10 @@ test("applySet", () => {
           "rootContext": [Circular],
           "target": P {
             "$": Object {
+              "$modelId": "id-1",
               "ch": P {
                 "$": Object {
+                  "$modelId": "id-2",
                   "ch": undefined,
                   "obj": undefined,
                   "y": 1,