+++ packages/lib/test/snapshot/fromSnapshot.test.ts
@@ -16,6 +16,7 @@ test("basic", () => {
   expect(p).toMatchInlineSnapshot(`
     P {
       "$": Object {
+        "$modelId": "id-2",
         "arr": Array [
           1,
           2,
@@ -23,6 +24,7 @@ test("basic", () => {
         ],
         "p2": P2 {
           "$": Object {
+            "$modelId": "id-1",
             "y": 12,
           },
           "$modelType": "P2",